+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Change Log - Power Platform Extension
 
+## 2.0.104
+- pac CLI 1.49.3, (see release notes on [nuget.org](https://www.nuget.org/packages/Microsoft.PowerApps.CLI/))
+
+## 2.0.102
+- Rollout of "Open in Desktop" capability in Power Pages Actions (Web)
+- Rollout of CodeQL security analysis support for Power Pages sites (Desktop)
+- Bug Fixes
+  - Fixed authentication button issue for Actions Hub in GitHub Codespaces
+  - Enhanced duplicate webpage handling in VS Code web
+
+## 2.0.100
+- pac CLI 1.48.2, (see release notes on [nuget.org](https://www.nuget.org/packages/Microsoft.PowerApps.CLI/))
+- Bug Fixes
+  - Fixed auth mismatch issue for Power Pages Actions (Desktop)
+  - Enabled Power Pages Actions(Desktop) in GitHub Codespaces.
+  - Added theme support for Power Pages Actions (Web) icons.
+
 ## 2.0.94
 - pac CLI 1.47.1, (see release notes on [nuget.org](https://www.nuget.org/packages/Microsoft.PowerApps.CLI/))
 - Bug Fixes